backend/requirements.txt
flask
flask-cors
pymysql
bcrypt
bleach
python-dotenv
requests
flask-sqlalchemy

